requirements.txt
MetaTrader5
pymongo
pandas
python-dotenv
pytz